+++ NEWS.md
@@ -1,3 +1,9 @@
+# dyngen 1.0.3
+
+* BUG FIX `generate_experiment()`: Return timepoint groups for `experiment_synchronised()`.
+
+* BUG FIX unit tests: loosen generate partitions constraints.
+
 # dyngen 1.0.2
 
 * MINOR CHANGE `generate_dataset()`: Fix subplot title.